move initializers, postInitializers, scriptsQueue, etc. to initialize() of Runtime Super Class
