+++ pyscriptjs/src/components/pyconfig.ts
@@ -4,15 +4,16 @@ import { appConfig } from '../stores';
 import type { Runtime, AppConfig } from '../runtime';
 import { PyodideRuntime } from '../pyodide';
 
-/*
-Usage of initializers, postInitializers, scriptsQueue, etc.
-is moved to `pyodide.ts` along with the `PyodideRuntime` class
-which now extends from the `Runtime` parent class. This is because
-all code pertaining to the pyodide runtime is in one place now,
-not polluting / mixing up with the functionality of `PyConfig`
-*/
-
 const DEFAULT_RUNTIME: Runtime = new PyodideRuntime();
+
+/**
+ * Configures general metadata about the PyScript application such
+ * as a list of runtimes, name, version, closing the loader
+ * automatically, etc.
+ *
+ * Also initializes the different runtimes passed. If no runtime is passed,
+ * the default runtime based on Pyodide is used.
+ */
 
 export class PyConfig extends BaseEvalElement {
     shadow: ShadowRoot;
